+++ CHANGELOG.md
@@ -5,25 +5,121 @@
 * AADConditionalAccessPolicy
   * Filter of CAPs by name with paging doesn't correctly filter, creating dupe CAPs, add more tests
     FIXES [#7350](https://github.com/Microsoft365DSC/Microsoft365DSC/issues/7350)
+* AADEntitlementManagementAccessPackage
+  * Fixed an issue where a resource role scope whose `AccessPackageResourceOriginId`
+    is specified as the object GUID was removed and re-added on every Set,
+    eventually leaving the access package with no resource roles. The desired
+    OriginId is now resolved to the same value that Get-TargetResource returns
+    before comparison.
+    FIXES [#7386](https://github.com/Microsoft365DSC/Microsoft365DSC/issues/7386)
+* AADApplicationFederatedIdentityCredential
+  * Added resource to manage federated identity credentials for Azure AD applications.
+* M365DSCConnection
+  * Fixed an issue with client secret authentication.
+* M365DSCReverse
+  * Fixed an issue where filtering was not applied correctly.
+
+# 1.26.729.2
+
+* M365DSCConnection
+  * Fixed issue connecting to a workload using `New-M365DSCConnection`.
+
+# 1.26.729.1
+
+* AADApplication
+  * Added new properties `DefaultRedirectUri`, `Info`, `Logo` and
+    `ServiceManagementReference`.
+    FIXES [#4321](https://github.com/Microsoft365DSC/Microsoft365DSC/issues/4321)
+* AADConditionalAccessPolicy
+  * Added requirement to specify `ExcludeExternalTenantsMembershipKind` if
+    `ExcludeExternalTenantsMembers` is specified during apply.
+  * Assigned the value `All` as default for `ExcludeExternalTenantsMembershipKind`
+    during apply.
+    FIXES [#6163](https://github.com/Microsoft365DSC/Microsoft365DSC/issues/6163)
+* AADEntitlementManagementAccessPackageAssignmentPolicy
+  * Fixed an issue where lookup by `DisplayName` returned all results.
+    FIXES [#7374](https://github.com/Microsoft365DSC/Microsoft365DSC/issues/7374)
+* AADEntitlementManagementAccessPackageCatalogResource
+  * Fixed an issue where a non-existing Service Principal or Group would
+    throw an error during the export instead of continuing.
+* EXOManagementRole
+  * Added retry and logging when `Get-ManagementRole` returns no results.
+* IntuneAppConfigurationPolicy
+  * Added information about no support for Settings catalog policy entries due to
+    delegated authentication requirements.
+    FIXES [#5672](https://github.com/Microsoft365DSC/Microsoft365DSC/issues/5672)
+* IntuneDeviceManagementDeviceDiagnosticSettings
+  * Initial release.
+    FIXES [#4154](https://github.com/Microsoft365DSC/Microsoft365DSC/issues/4154)
+* IntuneWindowsDataProcessingSettings
+  * Initial release.
+    FIXES [#4154](https://github.com/Microsoft365DSC/Microsoft365DSC/issues/4154)
+* M365DSCConnection
+  * Fixed an issue where `ApplicationSecret` was not converted to a secure string.
+* SentinelAlertRule
+  * Added `SubscriptionId` parameter to `Export-TargetResource`.
+  * Fixed an issue where non-Sentinel Log Analytics Workspaces were exported.
+* SentinelSetting
+  * Added `SubscriptionId` parameter to `Export-TargetResource`.
+  * Fixed an issue where non-Sentinel Log Analytics Workspaces were exported.
+* SentinelThreatIntelligenceIndicator
+  * Added `SubscriptionId` parameter to `Export-TargetResource`.
+  * Fixed an issue where non-Sentinel Log Analytics Workspaces were exported.
+* SentinelWatchlist
+  * Added `SubscriptionId` parameter to `Export-TargetResource`.
+  * Fixed an issue where non-Sentinel Log Analytics Workspaces were exported.
+* M365DSCDocGenerator
+  * Fixed an issue where not all permissions were shown on the resource docs page.
+* M365DSCExportUtil
+  * Removed validation from `SubscriptionId` parameter.
+* M365DSCReverse
+  * Updated `SubscriptionId` handling to pass to all export functions that define it.
+* M365DSCUtil
+  * Added the function `Update-M365DSCAuthenticationConfiguration` to update a
+    configuration to use a different authentication mode and save it to a new file.
+    FIXES [#3205](https://github.com/Microsoft365DSC/Microsoft365DSC/issues/3205)
+* MISC
+  * Updated documentation for the `Update-M365DSCAzureAdApplication` function.
+    FIXES [#5399](https://github.com/Microsoft365DSC/Microsoft365DSC/issues/5399)
+  * Updated required permissions and documentation for Azure resources.
+    FIXES [#6960](https://github.com/Microsoft365DSC/Microsoft365DSC/issues/6960)
+  * Added support to run Docker release images with older Microsoft365DSC versions
+* DEPENDENCIES
+  * Updated `MSCloudLoginAssistant` to version 1.1.72.
+    FIXES [#7371](https://github.com/Microsoft365DSC/Microsoft365DSC/issues/7371)
+
+# 1.26.722.1
+
 * O365OrgCustomizationSetting
   * Removed unused backtick in hashtable.
 * O365OrgSettings
   * Added missing workload connections.
 * O365SearchAndIntelligenceConfigurations
   * Added missing workload connection.
+* SCDLPComplianceRule
+  * Fixed an issue where trainable classifier ids in `AdvancedRule` could cause drift.
+    FIXES [#7352](https://github.com/Microsoft365DSC/Microsoft365DSC/issues/7352)
 * M365DSCExportUtil
+  * Added auto completion for `-Components` in `Export-M365DSCConfiguration`.
   * Since `M365DSCStringReplacementMap` is now being sent in parallel exports
     as global variable we need to set it as script scope based to be able to
-    inspect it
+    inspect it.
 * M365DSCReverse
   * Fixed an issue where parallel exports for EXO, O365 and SC could fail
     if combined with resources that use Microsoft Graph.
   * Fixed an issue where parallel exports were not being sent with the variable
     `M365DSCStringReplacementMap` and therefore mappings were not working and
-    creating drifts
+    creating drifts.
+  * Reduced the number of calls to `Get-M365DSCAllResources` if `-Components`
+    was specified.
+  * Reduced the number of calls to `Get-M365DSCAllResources` by caching its
+    results.
+* M365DSCUtil
+  * Added new helper function `Get-M365DSCAllResourcesPath`
 * DEPENDENCIES
-  * Updated `MSCloudLoginAssistant` to version 1.1.70.
+  * Updated `MSCloudLoginAssistant` to version 1.1.71.
 * MISC
+  * Reduced time to export Intune policies by up to 50%.
   * Updated function help descriptions and added / removed parameters.
 
 # 1.26.715.1
@@ -95,6 +191,8 @@
   * Fixed issue where function `Get-CompareParameters` was not being exported
 * IntuneWifiConfigurationPolicyIOS
   * Added `wpa3Personal` to the `WifiSecurityType` property.
+* SCDLPComplianceRule
+  * Added support for the `EndpointDlpRestrictions` property.
 * SettingsCatalogHelper
   * Fixed an issue where complex administrative template names
     were not handled correctly.